requirements.txt
anyio==3.6.2
cffi==1.15.1
click==8.1.3
cryptography==39.0.0
Deprecated==1.2.13
fastapi==0.89.1
greenlet==2.0.1
h11==0.14.0
idna==3.4
jwcrypto==1.4.2
pycparser==2.21
pydantic==1.10.4
PyJWT==2.6.0
python-dotenv==0.21.0
python-jwt==4.0.0
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
wrapt==1.14.1
